renaming model files to services
